Update TestOrders.java
Regressive changes. 
* Added explicit casting to 'day' and 'month' fields under MyDate() constructor.

+++ src/com/acmeordersystem/testing/TestOrders.java
@@ -9,10 +9,10 @@ public class TestOrders {
 	 * @param args
 	 */
 	public static void main(String[] args) {
-		MyDate date1 = new MyDate(1,20,2008);
+		MyDate date1 = new MyDate((byte)1, (byte)20, 2008);
 		Order anvil = new Order(date1, 2000.00, "Wile E Coyote", "Anvil", 10);
 
-		MyDate date2 = new MyDate(4,10,2008);
+		MyDate date2 = new MyDate((byte)4, (byte)10, 2008);
 		Order balloons = new Order(date2, 1000.00, "Bugs Bunny", "Balloon", 125);
 
 		// MyDate date3 = new MyDate(5, 20, 2008);
